+++ energy-prices/CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to the Energy Prices add-on will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.5.4] - 2026-01-01
+
+### Fixed
+- Corrected "Max Profit" calculation to use `Max Export - Min Import` instead of `Max Import - Min Import`
+- Updated MQTT sensor attributes for `max_profit_today` to show `min_import_price` and `max_export_price`
+
 ## [1.5.3] - 2026-01-01
 
 ### Added